set 3.6 as a minimum supported python version for installing
By default AFNI distributes python code in abin, a directory containing
all binaries and scripts. This change does not effect that.  It is
simply to drop the burden of maintainence for the cmake build system and
the circleci testing infrastructure (for example resolving testing
dependencies pre python 3.6 starts to become very difficult).