improve the cmake-build's management of AFNI's python code
The logic for managing python code has been moved to the cmakelists file
in the python_scripts directory

Now by default:

The python executable used for the build is hopefully a
little more intuitive. OSX framework python is used as a last resort.
Preference is given to python interpreters located in an environment
(conda or virtualenv), the default minimum version is 3.6 (this can be
overridden to use any python 3). The desired outcome is that the python
version used should be the python currently found first on the PATH
(assuming it meets the requirements).

The AFNI python package is installed into the current python interpreter
in development mode during the build, and installed into the
interpreter's site-packages during installation.

The code is installed using setuptools. pip is not used, avoid having
it as a dependency of the build, especially useful if an interpreter is
ideal except for the fact that it is missing pip. The above installation
described is equivalent to "pip install -e ." and "pip install ."

The build target "pytest" now checks that the python currently found and
used for tests does not differ from the one that was used during the
cmake configuration. If the interpreter does change by accident (due to
an accidental change in PATH/environment, or because the first python
found on the PATH does not satisfy the requirements) it can result in an
irritating state where dependencies are installed into the "wrong"
interpreter. For the brave or insistent this check can be removed by
setting FORCE_CURRENT_PY_INTERP_FOR_TESTS to "ON"