Finalize parallel test execution behavior
Went back and forth between xdist and pytest parallel trying to get
things to work. The main issues that were occurring were that there were
race conditions for xvfb displays, the physical display (on macos), and
data downloading via datalad. The latter two race conditions were
resolved using a resource lock for the process. This works nicely,
except that xdist uses more than one process and it fails.

Although the file locking implemented by xvfbwrapper is an approach that
is robust for both the xdist and the parallel pytest plugins, this may
cause issues with performance due to disk IO and the verbose output of
xdist is pretty unpleasant for a high number of cores (and this output
is inseperable from the verbose output of  pytest in general).

End result: xdist is not forbidden but a warning is issued and it fails
pretty badly (although iteratively succeeds) for data download. The
pytest-parallel plugin is now used by default and the freezing issue
that was occurring due to an incorrectly implemented lock for the
datalad download process has been resolved.