tidy component management for cmake build
rename components to have a more homogenous naming scheme.
corelibs is now a component.

add functionality to install a restricted list of components. For
example, just install the gui targets... this is helpful for packaging
into components but is not flawless. For example all "external
dependencies" would have to use system dependencies or build them first
(or possibly use the cmake superbuild pattern to fully control the build
and installation prior to population of the afni projects targets)

docs, funstuff, and readme now installed along with gui as configurable
installation components.

python is now set as required for the build. 3.6 is now the oldest python
version supported by the python foundation. Using python for adding
functionality to the build management instead of cmake would like be easier.