start to convert mri_mini to emcc compatible
the emscriptem compiler, emcc, has various details that need to be
addressed for the mri_mini target to work correctly. By default
functions are optimized away if not explicitly used and different
compile flags are used.

This commit is the beginings of cleanly isolating the mri_mini target
from the rest of the build so that downstreams can drive the afni build
conveniently to get access to the AFNI::mri_mini target without
inheritting any unnecessary dependencies, and using the  modern cmake
approach, all compile time definitions/link options etc. are all handled
in the afni build.